Moving label mapping types to types.h
