+++ src/search/cost_saturation/explicit_abstraction.h
@@ -2,10 +2,9 @@
 #define COST_SATURATION_EXPLICIT_ABSTRACTION_H
 
 #include "abstraction.h"
+#include "types.h"
 
 #include "../algorithms/priority_queues.h"
-#include "../algorithms/segmented_array_pool.h"
-#include "parallel_hashmap/phmap.h"
 
 #include <memory>
 #include <utility>
@@ -32,31 +31,6 @@ struct Successor {
 
 std::ostream &operator<<(std::ostream &os, const Successor &successor);
 
-using OpsPool = segmented_array_pool_template::ArrayPool<int>;
-using OpsSlice = segmented_array_pool_template::ArrayPoolSlice<int>;
-
-struct OpsSliceHash {
-    std::size_t operator()(OpsSlice v) const {
-        std::size_t seed = v.size();
-        for (int i : v) {
-            seed ^= std::hash<int>{}(i) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-        }
-        return seed;
-    }
-};
-
-struct OpsSliceEqualTo {
-    bool operator()(OpsSlice lhs, OpsSlice rhs) const {
-        if (lhs.size() != rhs.size()) {
-            return false;
-        }
-
-        return std::equal(lhs.begin(), lhs.end(), rhs.begin(), rhs.end());
-    }
-};
-
-using OpsToLabelId = phmap::flat_hash_map<OpsSlice, int, OpsSliceHash, OpsSliceEqualTo>;
-using LabelIdToOps = phmap::flat_hash_map<int, OpsSlice>;
 class ExplicitAbstraction : public Abstraction {
     int num_non_label_transitions;
     int num_label_transitions;
