+++ src/search/cost_saturation/types.h
@@ -1,6 +1,10 @@
 #ifndef COST_SATURATION_TYPES_H
 #define COST_SATURATION_TYPES_H
 
+#include "parallel_hashmap/phmap.h"
+
+#include "../algorithms/segmented_array_pool.h"
+
 #include <functional>
 #include <limits>
 #include <memory>
@@ -27,6 +31,31 @@ using CPFunction = std::function<CostPartitioningHeuristic(
 using CPHeuristics = std::vector<CostPartitioningHeuristic>;
 using DeadEnds = partial_state_tree::PartialStateTree;
 using Order = std::vector<int>;
+using OpsPool = segmented_array_pool_template::ArrayPool<int>;
+using OpsSlice = segmented_array_pool_template::ArrayPoolSlice<int>;
+
+struct OpsSliceHash {
+    std::size_t operator()(OpsSlice v) const {
+        std::size_t seed = v.size();
+        for (int i : v) {
+            seed ^= std::hash<int>{}(i) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+        }
+        return seed;
+    }
+};
+
+struct OpsSliceEqualTo {
+    bool operator()(OpsSlice lhs, OpsSlice rhs) const {
+        if (lhs.size() != rhs.size()) {
+            return false;
+        }
+
+        return std::equal(lhs.begin(), lhs.end(), rhs.begin(), rhs.end());
+    }
+};
+
+using OpsToLabelId = phmap::flat_hash_map<OpsSlice, int, OpsSliceHash, OpsSliceEqualTo>;
+using LabelIdToOps = phmap::flat_hash_map<int, OpsSlice>;
 }
 
 #endif
